+++ test/eca/test_helper.clj
@@ -60,6 +60,7 @@
 
 (defn reset-components! []
   (reset! config/initialization-config* {})
+  (reset! config/plugin-components* nil)
   (reset! components* (make-components))
   ;; Set default workspace folder for tests
   (swap! (db*) assoc :workspace-folders [{:uri (shared/filename->uri (System/getProperty "user.dir"))}]))